# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API-сервер для управления заявками и факторами риска в системе Msghistory.",
        "title": "API для системы Msghistory",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8090",
    "basePath": "/api",
    "paths": {
...
        "/msghistory": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает отфильтрованный список всех сформированных заявок (кроме черновиков и удаленных).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "msghistory"
                ],
                "summary": "Получить список заявок (авторизованный пользователь)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Фильтр по статусу заявки",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по дате 'от' (формат YYYY-MM-DD)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по дате 'до' (формат YYYY-MM-DD)",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ds.MsghistoryDTO"
                            }
                        }
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/msghistory/channelscart": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает ID черновика текущего пользователя и количество каналов в нем.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "msghistory"
                ],
                "summary": "Получить информацию для иконки корзины (авторизованный пользователь)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.CartBadgeDTO"
                        }
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/msghistory/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает полную информацию о заявке, включая привязанные каналы.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "msghistory"
                ],
                "summary": "Получить одну заявку по ID (авторизованный пользователь)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.MsghistoryDTO"
                        }
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Заявка не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Позволяет пользователю обновить поля своей заявки (возраст, пол, вес, рост).",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "msghistory"
                ],
                "summary": "Обновить данные заявки (авторизованный пользователь)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления",
                        "name": "updateData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.MsghistoryUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Логически удаляет заявку, переводя ее в статус \"удалена\".",
                "tags": [
                    "msghistory"
                ],
                "summary": "Удалить заявку (авторизованный пользователь)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/msghistory/{id}/channels/{channel_id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Изменяет дополнительное описание для конкретного канала в рамках одной заявки.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "m-m"
                ],
                "summary": "Обновить описание канала в заявке (авторизованный пользователь)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID канала",
                        "name": "channel_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новое описание",
                        "name": "updateData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.ChannelToMsghistoryUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет связь между заявкой и каналом.",
                "tags": [
                    "m-m"
                ],
                "summary": "Удалить канал из заявки (авторизованный пользователь)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID канала",
                        "name": "channel_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/msghistory/{id}/form": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Переводит заявку из статуса \"черновик\" в \"сформирована\".",
                "tags": [
                    "msghistory"
                ],
                "summary": "Сформировать заявку (авторизованный пользователь)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки (черновика)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Не все поля заполнены",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/msghistory/{id}/resolve": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Модератор завершает (с расчетом) или отклоняет заявку.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "msghistory"
                ],
                "summary": "Завершить или отклонить заявку (только модератор)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Действие: 'complete' или 'reject'",
                        "name": "action",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.MsghistoryResolveRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ds.CartBadgeDTO": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "msghistory_id": {
                    "type": "integer"
                }
            }
        },
...
        "ds.ChannelInMsghistoryDTO": {
            "type": "object",
            "properties": {
                "channel_id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "repost_level": {
                    "type": "integer"
                },
                "subscribers": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
        "ds.ChannelToMsghistoryUpdateRequest": {
            "type": "object",
            "properties": {
                "repost_level": {
                    "type": "integer"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
...
        "ds.MsghistoryDTO": {
            "type": "object",
            "properties": {
                "channels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.ChannelInMsghistoryDTO"
                    }
                },
                "coefficient": {
                    "type": "number"
                },
                "complition_date": {
                    "type": "string"
                },
                "coverage": {
                    "type": "number"
                },
                "creation_date": {
                    "type": "string"
                },
                "creator_login": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "forming_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "moderator_login": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "ds.MsghistoryResolveRequest": {
            "type": "object",
            "required": [
                "action"
            ],
            "properties": {
                "action": {
                    "description": "\"complete\" | \"reject\"",
                    "type": "string"
                }
            }
        },
        "ds.MsghistoryUpdateRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}